LazyContext.jl

LazyContext makes true lazy evaluation easy. Here is a simple port of with in R.

julia> using LazyContext

julia> with(d::WithContext, e::WithContext) = begin
           d_evaluated = evaluate!(d)
           e_copy = copy(e)
           merge!(e_copy.environment, d_evaluated)
           evaluate!(e_copy)
       end;

julia> @new_environment;

julia> @import_to_environment with;

julia> @evaluate begin
           d = Dict(:a => 1, :b => 2)
           @with d a + b
       end
3

Merging will work for any type which can be key-value iterated.

LazyContext.Environment

LazyContext.WithContext

LazyContext.copy_of_base_environment

LazyContext.evaluate!

LazyContext.immutable_merge

LazyContext.@evaluate

LazyContext.@import_to_environment

LazyContext.@new_environment

LazyContext.@use_in_environment

LazyContext.Environment — Type.

type Environment
    dictionary::Base.ImmutableDict{Symbol,Any}()
end

A type to store all the names in your local environment.

Can access and record new values in an environment.

julia> using LazyContext

julia> e = Environment();

julia> e[:a] = 1;

julia> e
Environment. Most recent assignment: a

julia> e[:a]
1

By default, contains an :ENVIRONMENT key which points to itself. If the environment has a parent, it also will contain a :PARENT key which points to its parent.

julia> using LazyContext

julia> e = Environment();

julia> e[:a] = 1;

julia> e[:ENVIRONMENT][:a]
1

julia> e2 = copy(e);

julia> e2[:PARENT][:a]
1

julia> e2[:b] = 1;

julia> e2[:PARENT][:b]
ERROR: KeyError: key :b not found
[...]

Can merge! or merge new objects into an environment. You can merge in modules:

julia> using LazyContext

julia> e = Environment();

julia> merge!(e, Core);

julia> e[:Symbol]
Symbol

and dictonaries:

julia> using LazyContext

julia> e = Environment();

julia> merge!(e, Dict(:a => 1) );

julia> e[:a]
1

Extend immutable_merge to be able to merge in more types.

source

LazyContext.WithContext — Type.

type WithContext
    expression::Any
    environment::Environment
end

An expression in the context of its Environment

source

LazyContext.copy_of_base_environment — Method.

copy_of_base_environment()

Get an environment which contains Core and Base. Has no parent.

source

LazyContext.evaluate! — Method.

evaluate!(w; locked = false)

Evaluate w.expression in context of w.environment.

Unlocked environments cannot be directly accessed in a type stable way. If you include the locked = true keyword, however, the bindings in the environment will be fixed allowing for type-stable access.

julia> using LazyContext

julia> @new_environment;

julia> ENVIRONMENT[:a] = 1;

julia> ENVIRONMENT[:b] = 2;

julia> w = WithContext(:(a + b), ENVIRONMENT);

julia> evaluate!(w)
3

julia> evaluate!(w, locked = true)
3

julia> evaluate!( WithContext(:(a = 1), ENVIRONMENT), locked = true)
ERROR: Cannot assign 1 to a: environment is locked
[...]

source

LazyContext.immutable_merge — Method.

immutable_merge(d, a)

Merge d (usually an ImmutableDict) and a into a new ImmutableDict. For example:

julia> using LazyContext

julia> d = Base.ImmutableDict{Symbol, Any}();

julia> a = Dict(:b => 2);

julia> immutable_merge(d, a)[:b]
2

source

LazyContext.@evaluate — Macro.

macro evaluate(expression)

evaluate! expression in the context of ENVIRONMENT.

Only limited syntax is supported inside @evaluate.

Anonymous are suppported, but only using the terse syntax.

julia> using LazyContext

julia> @new_environment;

julia> @evaluate begin
           test_function = (a, b...; c = 1) -> +(a, c, b...)
           test_function(1, 2, 3; c = 4)
       end
10

Anonymous functions create new environments. The parent environment can be accessed with PARENT, while the local environment can be accessed with ENVIRONMENT.

julia> using LazyContext

julia> @new_environment;

julia> @evaluate begin
           a = 1
           test = b -> begin
               merge!(ENVIRONMENT, Dict(:a => 2))
               a + b + PARENT[:a]
           end
           test(3)
       end
6

Anonymous functions use dynamic scoping: changes in global variables will change functions which reference them.

julia> using LazyContext

julia> @new_environment;

julia> (3, 4) == @evaluate begin
           a = 1
           test = b -> a + b
           first_result = test(2)
           a = 2
           second_result = test(2)
           first_result, second_result
       end
true

Macro calls do not call macros. Instead, they create WithContext expression wrappers out of all arguments and pass them to the function with the same name.

julia> using LazyContext

julia> @new_environment;

julia> e = @evaluate begin
           a = 1
           b = 2
           @identity a + b
       end
WithContext(a + b)

julia> evaluate!(e)
3

source

LazyContext.@import_to_environment — Macro.

macro import_to_environment(args...)

Make args, a list of names, available in the global ENVIRONMENT.

julia> using LazyContext

julia> @new_environment;

julia> import MacroTools;

julia> @import_to_environment LazyContext MacroTools;

julia> Environment == @evaluate LazyContext.Environment
true

julia> MacroTools.prettify == @evaluate MacroTools.prettify
true

source

LazyContext.@new_environment — Macro.

macro new_environment

Will create a new copy_of_base_environment named ENVIRONMENT. Prerequisite for many macros in this package. Be careful not to overwrite ENVIRONMENT.

julia> using LazyContext

julia> @new_environment;

julia> Dict == ENVIRONMENT[:Dict]
true

source

LazyContext.@use_in_environment — Macro.

use_in_environmnet(args...)

Make the contents of each item in args available in the global ENVIRONMENT, by merging in each one by one.

julia> using LazyContext

julia> @new_environment;

julia> import MacroTools;

julia> @use_in_environment LazyContext MacroTools;

julia> Environment == @evaluate Environment
true

julia> MacroTools.prettify == @evaluate prettify
true

source